feat(component): extract agnostic Component, wire React as adapter
Distill the renderer-agnostic essence of Component out of @expressive/react
into the transitional @expressive/component package. React becomes a thin
adapter that completes the class without inversion-of-control seams.

/component owns the neutral core:
- class skeleton, props reconciliation, StrictMode dedup (PENDING)
- recovery: on a double-construct, arms a one-shot restore closure under
  Symbol.for('React.StrictMode') capturing the live reactive accessors
- subcomponent discovery via exported discover()/Realize

/react completes it as an adapter:
- augments Host.node; attaches bootstrap/contextType/compat shims/ErrorBoundary
  to the prototype
- wires React-internal field claim + subcomponent realization via Component.on
- invokes the revive closure from its context setter (the remount trigger)

Removes the realize/onConstruct module seams and the restore export. The
adapter-core boundary is now type augmentation + Component.on + discover + a
Symbol handoff - no exported recovery function, no @expressive/state change.
