+++ packages/react/package.json
@@ -45,6 +45,7 @@
     "preversion": "npm run test && npm run build"
   },
   "dependencies": {
+    "@expressive/component": "workspace:^",
     "@expressive/state": "workspace:^"
   },
   "devDependencies": {
